+++ packages/adminui-vue/components/controls/mixins/selection/logic.js
@@ -31,27 +31,13 @@ module.exports = {
     }
   },
   methods: {
-    // WARNING: this is to open the edit form with one click when the row is already selected
-    focusHandler (rowIndex) {
-      if (this.hoverIndex === rowIndex) return
-      setTimeout(() => {
-        this.hoverIndex = rowIndex
-      }, 150)
-    },
-    // used in UCardView
-    cardClickHandler (rowIndex, event) {
-      // const row = this.items[rowIndex]
-      // the order in which the methods are called is important
-      // this.handlerSelection(row, event)
-      this.onTableRowClickHandler(rowIndex, event)
-    },
-    onInputClickHandler (row, event) {
+    onMultiSelectionClick (rowIndex, event) {
+      const row = this.items[rowIndex]
       this.handlerSelection(row, event)
-      this.$emit('click', { row })
-      this.$emit('click-row', { row })
+      this.onTableRowClickHandler(rowIndex)
     },
     async handlerSelection (row, event) {
-      this.$emit('click-cell', { row })
+      this.$emit('click-cell', { row, isMultipleSelectionCell: true })
       if (!this.enableMultiSelect) return
       const { multiSelectKeyAttr, curSelection } = this
       const arr = curSelection
@@ -95,7 +81,7 @@ module.exports = {
       /**
        * Triggers when the user removes or adds a selection
        *
-       * @param {array<number>} selectedIDs an array that includes all currently selected values
+       * @param {Array<number>} selectedIDs an array that includes all currently selected values
        */
       this.$emit('selected', this.curSelection)
     },
@@ -105,7 +91,7 @@ module.exports = {
       const temp = new Set(this.curSelection)
       if (!allSelected) {
         const addedCache = []
-        items.forEach(el => {
+        items.forEach((el) => {
           const value = el[multiSelectKeyAttr]
           if (!temp.has(value)) {
             temp.add(value)
@@ -123,18 +109,18 @@ module.exports = {
       }
       this.emitSelectionEvent()
     },
-    async contextMenuEventHandler ($event, row, col) {
+    async contextMenuEventHandler ($event, rowIndex, col) {
+      const row = this.items[rowIndex]
+      this.setCurrentRow(rowIndex)
       if (this.enableMultiSelect) {
         const { multiSelectKeyAttr, curSelection } = this
         const value = row[multiSelectKeyAttr]
         if (!curSelection.includes(value)) {
           const cache = this.getSelectionRows()
-          for (const row of cache) {
-            await this.handlerSelection(row)
+          for (const rowInCache of cache) {
+            await this.handlerSelection(rowInCache, $event)
           }
-          const rowIndex = this.items.indexOf(row)
-          this.setCurrentRow(rowIndex)
-          await this.handlerSelection(row)
+          await this.handlerSelection(row, $event)
         }
       }
       /**
@@ -147,13 +133,13 @@ module.exports = {
        *
        * @deprecated - exist for backward copability. Use contextmenu menu event and check column available field
        */
-      if (col) this.$emit('contextmenu-cell', { event: $event, row, column: col })
+      if (col) { this.$emit('contextmenu-cell', { event: $event, row, column: col }) }
     },
     getSelectionRows () {
       const { items, multiSelectKeyAttr } = this
       const rows = []
       const temp = new Set(this.curSelection)
-      items.forEach(el => {
+      items.forEach((el) => {
         const value = el[multiSelectKeyAttr]
         if (temp.has(value)) {
           temp.delete(value)
@@ -269,7 +255,8 @@ module.exports = {
       // because first tr in table it is head
       const shift = children.length - this.items.length
       const row = children[this.hoverIndex + shift]
-      if (!row || document.activeElement === row) return
+      if (!row) return
+      // needed to navigating with arrow
       row.focus()
     }
   }